hatch: Actually include sourcefiles in wheel
After I published v0.1.7, I noticed the source files were missing from the wheel.
This makes that version of the library impossible to use.

I have yanked v0.1.7 and will publish this commit as v0.1.8.
The only change is the fixed wheel.

This was caused by the transition from setuptools -> hatch
and it should not happen again.